Fix #136 Add an example image classification notebook
