--- Data_Collection/images.json
@@ -1,109 +1,0 @@
-{
-    "Records": [
-	{
-            "keywords": "washing machine"
-        },
-        {
-            "keywords": "telephone"
-        },
-        {
-            "keywords": "sewing machine"
-        },
-        {
-            "keywords": "radio"
-        },
-        {
-            "keywords": "stereo"
-        },
-        {
-            "keywords": "sofa"
-        },
-        {
-            "keywords": "ceiling fan"
-        },
-        {
-            "keywords": "table fan"
-        },
-        {
-            "keywords": "armchair"
-        },
-        {
-            "keywords": "cattle"
-        },
-        {
-            "keywords": "mattress"
-        },
-        {
-            "keywords": "chair"
-        },
-        {
-            "keywords": "tube light"
-        },
-        {
-            "keywords": "bulb light"
-        },
-        {
-            "keywords": "gas cylinder"
-        },
-        {
-            "keywords": "stove"
-        },
-        {
-            "keywords": "electric iron"
-        },
-        {
-            "keywords": "kitchen sink"
-        },
-        {
-            "keywords": "desktop computer"
-        },
-        {
-            "keywords": "food mixer"
-        },
-        {
-            "keywords": "food processor"
-        },
-        {
-            "keywords": "pressure cooker"
-        },
-        {
-            "keywords": "cars"
-        },
-        {
-            "keywords": "bicycles"
-        },
-        {
-            "keywords": "motorcycles"
-        },
-        {
-            "keywords": "vacuum cleaner"
-        },
-        {
-            "keywords": "thermos flask"
-        },
-        {
-            "keywords": "dish antenna"
-        },
-        {
-            "keywords": "firewood"
-        },
-        {
-            "keywords": "brick wall"
-        },
-        {
-            "keywords": "stone wall"
-        },
-        {
-            "keywords": "mud walls"
-        },
-        {
-            "keywords": "wardrobe"
-        },
-        {
-            "keywords": "oven"
-        },
-        {
-            "keywords": "microwave oven"
-        }
-    ]
-}
